+++ sender-entrypoint.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+set -e
+
+# run migrations
+alembic upgrade head
+
+# print and run image cmd
+echo "$@"
+exec "$@"