+++ src/sender/db.py
@@ -1,4 +1,7 @@
-from sqlalchemy import create_engine
+import logging
+
+from sqlalchemy import create_engine, text
+from sqlalchemy.exc import OperationalError
 from sqlalchemy.orm import Session, sessionmaker
 
 from src.sender.settings import Settings
@@ -7,21 +10,21 @@ from src.sender.settings import Settings
 class Database:
     def __init__(self, settings: Settings):
         self.settings = settings
-
-        if settings.DB_TYPE == "mysql":
-            self.db_url = (
-                f"mysql+pymysql://{settings.DB_USER}:{settings.DB_PASSWORD}@"
-                f"{settings.DB_HOST}:{settings.DB_PORT}/{settings.DB_DATABASE}"
-            )
-        elif settings.DB_TYPE == "postgresql":
-            self.db_url = (
-                f"postgresql+psycopg2://{settings.DB_USER}:{settings.DB_PASSWORD}@"
-                f"{settings.DB_HOST}:{settings.DB_PORT}/{settings.DB_DATABASE}"
-            )
-        else:
-            raise ValueError(f"Unsupported DB_TYPE: {settings.DB_TYPE}")
+        self.db_url = settings.DATABASE_URL
 
         self.engine = create_engine(self.db_url, echo=False)
+        try:
+            with self.engine.connect() as connection:
+                connection.execute(text("SELECT 1;"))
+                logging.info("Database connection successful! 🎉")
+
+        except OperationalError as e:
+            logging.error("Error: Could not connect to the database. Details: %s", e)
+            raise e
+        except Exception as e:
+            logging.error("An unexpected error occurred: %s", e)
+            raise e
+
         self.SessionLocal = sessionmaker(bind=self.engine)
 
     def get_session(self) -> Session: