+++ src/sender/entities/user.py
@@ -6,15 +6,16 @@ from src.sender.entities.base import Base
 class User(Base):
     __tablename__ = "users"
     id = Column(Integer, primary_key=True, autoincrement=True)
-    token = Column(String, nullable=False, unique=True)
-    name = Column(String, nullable=False, unique=True)
-    organization = Column(String, nullable=True)
-    email = Column(String, nullable=True)
+    token = Column(String(length=255), nullable=False, unique=True)
+    name = Column(String(length=255), nullable=False, unique=True)
+    organization = Column(String(length=255), nullable=True)
+    email = Column(String(length=255), nullable=True)
     priority = Column(Integer, nullable=False)
     threshold = Column(Integer, nullable=False)
-    client_type = Column(String)
+    client_type = Column(String(length=255))
     default_routing_mode = Column(
         Enum("any", "private-first", "private-only", native_enum=False),
         nullable=False,
+        server_default="any",
         default="any",
     )